+++ freecad/spreadsheetplus/commands/__init__.py
@@ -1,8 +1,6 @@
 # SPDX-License-Identifier: LGPL-2.1-or-later
 """SpreadSheetPlus commands."""
 
-from .create_sheet import CreateSheet as CreateSheet
 from .create_master_sheet import CreateMasterSheet as CreateMasterSheet
 from .create_config_ref import CreateConfigRef as CreateConfigRef
-from .edit_config_table import EditConfigTable as EditConfigTable
 from .switch_configuration import SwitchConfiguration as SwitchConfiguration