+++ freecad/spreadsheetplus/dialogs/__init__.py
@@ -1,5 +1,4 @@
 # SPDX-License-Identifier: LGPL-2.1-or-later
 """SpreadSheetPlus dialogs."""
 
-from .table_editor import TableEditorDialog as TableEditorDialog
 from .select_configuration import SelectConfigurationDialog as SelectConfigurationDialog