+++ freecad/spreadsheetplus/resources/translations/SpreadSheetPlus_de.ts
@@ -12,14 +12,6 @@
         <translation>Eine Tabellenkalkulations-Workbench ähnlich der Standard-Workbench</translation>
     </message>
     <message>
-        <source>Create spreadsheet</source>
-        <translation>Tabelle erstellen</translation>
-    </message>
-    <message>
-        <source>Create a new spreadsheet sheet</source>
-        <translation>Eine neue Tabelle erstellen</translation>
-    </message>
-    <message>
         <source>Create MasterSheet</source>
         <translation>MasterSheet erstellen</translation>
     </message>
@@ -36,48 +28,12 @@
         <translation>Einen ConfigRef mit der ausgewählten Tabelle verknüpfen</translation>
     </message>
     <message>
-        <source>Edit configuration table</source>
-        <translation>Konfigurationstabelle bearbeiten</translation>
-    </message>
-    <message>
-        <source>Edit the selected spreadsheet&apos;s configuration table</source>
-        <translation>Die Konfigurationstabelle der ausgewählten Tabelle bearbeiten</translation>
-    </message>
-    <message>
         <source>Switch configuration</source>
         <translation>Konfiguration wechseln</translation>
     </message>
     <message>
         <source>Change the selected ConfigRef&apos;s configuration row</source>
         <translation>Die Konfigurationszeile des ausgewählten ConfigRef ändern</translation>
-    </message>
-    <message>
-        <source>Add configuration</source>
-        <translation>Konfiguration hinzufügen</translation>
-    </message>
-    <message>
-        <source>Remove configuration</source>
-        <translation>Konfiguration entfernen</translation>
-    </message>
-    <message>
-        <source>Add parameter</source>
-        <translation>Parameter hinzufügen</translation>
-    </message>
-    <message>
-        <source>Remove parameter</source>
-        <translation>Parameter entfernen</translation>
-    </message>
-    <message>
-        <source>Configuration name:</source>
-        <translation>Konfigurationsname:</translation>
-    </message>
-    <message>
-        <source>Parameter name:</source>
-        <translation>Parametername:</translation>
-    </message>
-    <message>
-        <source>Table OK</source>
-        <translation>Tabelle OK</translation>
     </message>
     <message>
         <source>Type to search…</source>